PO-2320 Use feature toggle annotation for central fund
Replace deprecated FeatureToggleService usage with the shared @FeatureToggle pattern and move disabled-flag coverage to integration tests.
