+++ src/main/java/uk/gov/hmcts/opal/controllers/CentralFundController.java
@@ -11,29 +11,29 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.opal.common.launchdarkly.FeatureDisabledException;
-import uk.gov.hmcts.opal.common.launchdarkly.service.FeatureToggleService;
+import uk.gov.hmcts.opal.common.launchdarkly.FeatureToggle;
 import uk.gov.hmcts.opal.dto.CentralFundResponse;
 import uk.gov.hmcts.opal.generated.model.GetCentralFundByBusinessUnit200Response;
 import uk.gov.hmcts.opal.service.CentralFundService;
+import uk.gov.hmcts.opal.util.FeatureFlags;
 
 @RestController
 @Slf4j(topic = "opal.CentralFundController")
 @RequiredArgsConstructor
 public class CentralFundController {
 
-    private static final String RELEASE_1B_FEATURE = "release-1b";
-
     private final CentralFundService centralFundService;
-    private final FeatureToggleService featureToggleService;
 
+    @FeatureToggle(
+        feature = FeatureFlags.RELEASE_1B,
+        defaultValueProperty = FeatureFlags.RELEASE_1B_ENABLED_PROPERTY
+    )
     @GetMapping(value = "/central-funds/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<GetCentralFundByBusinessUnit200Response> getCentralFundByBusinessUnit(
         @PathVariable("id") Integer businessUnitId,
         @RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String authHeaderValue) {
 
         log.debug(":GET:getCentralFundByBusinessUnit: businessUnitId={}", businessUnitId);
-        checkRelease1bEnabled();
 
         CentralFundResponse response = centralFundService.getCentralFundByBusinessUnit(
             businessUnitId,
@@ -43,13 +43,5 @@ public class CentralFundController {
         return ResponseEntity.ok()
             .eTag(createETag(response))
             .body(response.getPayload());
-    }
-
-    private void checkRelease1bEnabled() {
-        if (!featureToggleService.isFeatureEnabled(RELEASE_1B_FEATURE)) {
-            String message = "Feature release-1b is not enabled for method getCentralFundByBusinessUnit";
-            log.debug(message);
-            throw new FeatureDisabledException(message);
-        }
     }
 }
